game works now!

+++ src/main/java/org/example/Configuration.java
@@ -1,7 +1,7 @@
 package org.example;
 
 public class Configuration {
-    public static int BP;
+    public static double BP;
     public static int BD;
     public static int LD;
     public static int VMax;
@@ -12,8 +12,8 @@ public class Configuration {
     public final static int GRID_WIDTH = 1200;
     public final static int GRID_HEIGHT = 1200;
     public static int IW;
-    public static int IS;
-    public Configuration(int LethalDose, int MaximumSpeedLimit, int BonusLength, int BonusSpeed, int TailLength, int MapSize, int InitialLength, int InitialWidth, int InitialSpeed) {
+    public static double IS;
+    public Configuration(int LethalDose, int MaximumSpeedLimit, int BonusLength, double BonusSpeed, int TailLength, int MapSize, int InitialLength, int InitialWidth, double InitialSpeed) {
         LD = LethalDose;
         VMax = MaximumSpeedLimit;
         BD = BonusLength;
